+++ api/src/routes/index.js
@@ -19,6 +19,6 @@ router.use('/users', require('./users') /* #swagger.security = [{"BearerAuth": [
 router.use('/workflows', require('./workflows') /* #swagger.security = [{"BearerAuth": []}] */);
 router.use('/projects', require('./projects') /* #swagger.security = [{"BearerAuth": []}] */);
 router.use('/statistics', require('./statistics'));
-router.use('/ingest', require('./ingest'));
+router.use('/fs', require('./fs'));
 
 module.exports = router;
