--- ui/src/services/ingest.js
@@ -1,9 +1,0 @@
-import api from "./api";
-
-class IngestionService {
-  getPathFiles({ path, dirs_only }) {
-    return api.get("/ingest", { params: { path, dirs_only } });
-  }
-}
-
-export default new IngestionService();
